Bugfix rollup: Map recentering, missing par data, stale geolocations
- Map recentering
When loading a round from localStorage, if the OSM data
is not already cached, the map will fail to recenter. Fix is to add
a recentering explicitly.

- Missing par data
Not all courses have par data input. If no par data is added, just skip

- Stale geolocation check can be undefined
If checking geolocation before the browser returns a location, we can
get an object undefined error. Fix is to check if defined first.